feat: refactor canvas structure and update version to 0.0.4 in package files
